Add support for the ++help command
help messages are actually stored as a single string in PROGMEM
because it's a pain in the neck to store an array of strings there.

Support both with or without argument:
- without argument: display all help messages,
- with argument: displa the message for the command given as argument.

Each line is a help message for a command (aka '\n' char is used as
a separator between commands), so no support for multiline help message
for now.

For each line, the first ':' char is used as a separator between the
command name and the help message.